move PBT back to verifier tests;
and the rest to integration tests;
